requirements.txt
flask>=3.0.0
flask-bcrypt>=1.0.1
PyJWT>=2.8.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
pymysql>=1.1.0
cryptography>=42.0.0
feedparser>=6.0.10
python-dotenv>=1.0.0
gunicorn>=21.2.0
